Removes CampaignIdSingle code
