+++ resources/assets/components/Campaign/campaign.scss
